Move haskell analyzers to their own files
